feat: remove api & update pagination
